refactor(auth-shell): extract WorkOS provider into shared package
Moves AuthKitProvider mount, RootErrorBoundary, and useApi hook into
@us-rse/auth-shell so both the web SPA and the upcoming admin app can
share them without diverging. apps/web shims become thin re-exports.
