+++ apps/web/package.json
@@ -12,6 +12,7 @@
   },
   "dependencies": {
     "@tailwindcss/vite": "^4.2.4",
+    "@us-rse/auth-shell": "*",
     "@us-rse/design-system": "*",
     "@workos-inc/authkit-react": "^0.16.1",
     "maplibre-gl": "^5.24.0",
