+++ packages/auth-shell/src/index.ts
@@ -1,0 +1,3 @@
+export { AuthShell } from "./AuthShell";
+export { RootErrorBoundary } from "./RootErrorBoundary";
+export { useApi } from "./useApi";
